Add global strictMode config var (#3)
* Add global environment config var

Can be used to put the node in "development" or "production" mode.
Adds the testify package.

* Add env flag to PrintConfig

* Add IsProductionMode method to global config

Hels to determine if the node is running in production mode of not. If so, modules can perform checks on config settings.

* Renamed ProductionMode to StrictMode and made it a bool

Instead of naming deployments like production and developement, just talk about strict/unstrict. This flag itself does nothing but can be used for other modules to test settings. E.g. do not accepts IRMA demo attributes when strict mode is on.